+++ tools/tsp-client/src/npm.ts
@@ -19,9 +19,9 @@ export async function createPackageJson(rootPath: string, deps: Set<string>): Pr
   await writeFile(filePath, packageJson);
 }
 
-export function installDependencies(workingDir: string): Promise<void> {
+export async function npmCommand(workingDir: string, args: string[]): Promise<void> {
   return new Promise((resolve, reject) => {
-    const npm = spawn("npm", ["install", "--no-lock-file"], {
+    const npm = spawn("npm", args, {
       cwd: workingDir,
       stdio: "inherit",
       shell: true,
@@ -30,11 +30,11 @@ export function installDependencies(workingDir: string): Promise<void> {
       if (code === 0) {
         resolve();
       } else {
-        reject(new Error(`npm failed exited with code ${code}`));
+        reject(new Error(`npm ${args[0]} failed exited with code ${code}`));
       }
     });
     npm.once("error", (err) => {
-      reject(new Error(`npm install failed with error: ${err}`));
+      reject(new Error(`npm ${args[0]} failed with error: ${err}`));
     });
   });
 }